requirements.txt
flask
flask-cognito-auth
Whisper
pydub
nltk
boto3
python-dotenv
openai
supabase
Flask-Uploader